Refactor validator to validate schema with validate only
Add license to unit test

convert json factory schema to local var
