--- pinot-core/src/main/java/org/apache/pinot/core/util/validator/PinotConfigSchemasValidator.java
@@ -1,44 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.apache.pinot.core.util.validator;
-
-import com.github.fge.jsonschema.core.exceptions.ProcessingException;
-import java.io.IOException;
-
-/**
- * Factory to help create validator for [Table Config,]
- * */
-public class PinotConfigSchemasValidator extends BasePinotConfigValidator {
-
-
-  private static final String TABLE_CONFIG_SCHEMA = "/schemas/tableConfig.json";
-
-  private PinotConfigSchemasValidator(String schema)
-      throws IOException, ProcessingException {
-    loadSchema(schema);
-  }
-
-  /**
-   * Create validator for table config
-   * */
-  public static PinotConfigSchemasValidator forTableConfig()
-      throws IOException, ProcessingException {
-    return new PinotConfigSchemasValidator(TABLE_CONFIG_SCHEMA);
-  }
-}
