--- pinot-core/src/test/java/org/apache/pinot/util/validator/PinotConfigSchemasValidatorTest.java
@@ -1,9 +1,0 @@
-package org.apache.pinot.util.validator;
-
-import org.testng.annotations.Test;
-
-
-public class PinotConfigSchemasValidatorTest {
-
-
-}
